--- utf8-lua54/lutf8lib_lua54.h
@@ -1,15 +1,0 @@
-//
-//  lutf8lib_lua52.h
-//  Lua
-//
-//  Created by Robert Patterson on 6/3/22.
-//
-
-#ifndef lutf8lib_lua52_h
-#define lutf8lib_lua52_h
-
-#include "lua.h"
-
-LUAMOD_API int luaopen_utf8_standalone (lua_State *L);
-
-#endif /* lutf8lib_lua52_h */
